Move some env check/setting related codes from api.php to /core/env.php
